Converted from asynchronous dispatcher to synchronous event bus
Given that multithreading is really not paying off here, I will back off and move this change to a more traditional synchronous processing. While asynchronous can provide benefits, it is something that a system can manage on its own.

Instead of using Dispatcher to send events, I'll see if CDI can provide better support. So far, event listeners cannot be ordered; that may be a problem to solve for another day though.