Refactored Flow Candidates Calculation
The naive flow candidates considered all 4 million blocks.

Instead of naively checking every potential candidates, only consider blocks that may have introduced an equilibrium disturbance (including the flow system itself!)

For that, any change in water level or block placement introduces a potential distrubion. After getting all the flow candidates, consider the system in equilibrium: if any flow candidates flowed... We know we are not yet in a state of equilibrium (and we mark the block with movement as disturbed.)

This effectively narrows our search for flow candidates to active water regions rather than a comprehensive scan of the full grid.